add gravater, versions fixed + moved logic to ViewModel
