+++ Lab3/app/build.gradle
@@ -7,7 +7,8 @@ plugins {
 
 apply from: '../version.gradle'
 
-def version = getArray()
+def tag = getLastTag()
+def patch = getPatch()
 
 android {
     compileSdkVersion 30
@@ -17,8 +18,8 @@ android {
         applicationId "com.example.battleship"
         minSdkVersion 24
         targetSdkVersion 30
-        versionCode 1
-        versionName "${version[0]}.${version[1]}.${version[2]}"
+        versionCode generateVersionCode(tag , patch)
+        versionName "${tag}.${patch}"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -52,6 +53,7 @@ dependencies {
     implementation 'com.google.firebase:firebase-analytics:18.0.2'
     implementation 'com.google.firebase:firebase-database-ktx:19.6.0'
     implementation 'com.google.firebase:firebase-storage-ktx:19.2.1'
+    implementation files('libs\\gravatar-android-1.0.jar')
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
@@ -61,4 +63,6 @@ dependencies {
     kapt 'com.github.bumptech.glide:compiler:4.12.0'
     implementation 'androidx.appcompat:appcompat:1.2.0'
     implementation 'com.google.android.material:material:1.3.0'
+    implementation 'com.squareup.picasso:picasso:2.5.2'
+    implementation "android.arch.lifecycle:extensions:1.1.1"
 }
