+++ Lab3/app/src/main/res/values/strings.xml
@@ -6,5 +6,6 @@
         <item>Ghoul</item>
         <item>Wolf</item>
         <item>Pikachu</item>
+        <item>Gravatar</item>
     </string-array>
 </resources>
