+++ Lab3/version.gradle
@@ -1,35 +1,41 @@
-def getArray() {
+def getPatch() {
     def diffOutput = new ByteArrayOutputStream()
     exec {
-        commandLine  'cmd', 'git', 'describe',
-                standardOutput = diffOutput
+        commandLine 'cmd','git', 'describe'
+        standardOutput = diffOutput
     }
-    def ReceivedString = diffOutput.toString()
-    def currentDotPosition = ReceivedString.indexOf('.')
-    List VersionName = new ArrayList()
-    VersionName.add(ReceivedString.substring(1,currentDotPosition))
-    ReceivedString = ReceivedString.substring(currentDotPosition+1, ReceivedString.length())
-    currentDotPosition = ReceivedString.indexOf('-')
-    if (currentDotPosition!=-1){
-        VersionName.add(ReceivedString.substring(0,currentDotPosition))
-        ReceivedString = ReceivedString.substring(currentDotPosition+1, ReceivedString.length())
-        currentDotPosition = ReceivedString.indexOf('-')
-        VersionName.add(ReceivedString.substring(0, currentDotPosition))
+    def res = diffOutput.toString()
+    if (res.contains('-')) {
+        return res.split('-')[1]
+    } else {
+        return '0'
     }
-    else {
-        VersionName.add(ReceivedString.substring(0, ReceivedString.length()-1))
-        VersionName.add(0)
-    }
-    return VersionName
-
 }
 
-static def generateVersionCode(ArrayList array) {
-    def result = (array[0].toInteger()*10000 + array[1].toInteger())*100 + array[2].toInteger()
-    return result
+def getLastTag() {
+    def tagsOutput = new ByteArrayOutputStream()
+    exec {
+        commandLine 'cmd', 'git', 'describe'
+        standardOutput = tagsOutput
+    }
+    def res = tagsOutput.toString()
+    if (res.contains('-')) {
+        return res.split('-').first()
+    } else {
+        return '0.1'
+    }
+}
+
+static def generateVersionCode(String tag, String patch) {
+    def dot = tag.indexOf('.')
+    def major = tag.substring(0, dot).toInteger()
+    def minor = tag.substring(dot + 1).toInteger()
+    def res = (major*10 + minor.toInteger())*100 + patch.toInteger()
+    return res
 }
 
 ext {
-    getArray = this.&getArray
+    getPatch = this.&getPatch
+    getLastTag = this.&getLastTag
     generateVersionCode = this.&generateVersionCode
 }
